requirements.txt
attrdict==2.0.0
numpy==1.12.0
pandas==0.19.2
sh==1.12.9
scipy==0.18.1
cycler==0.10.0
tensorflow_gpu==1.0.0
matplotlib==2.0.0
tqdm==4.11.2
ipython==6.1.0
Pillow==4.2.1
seaborn==0.7.1
tensorflow==1.2.1
PyYAML==3.12
